requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
gunicorn==21.2.0
waitress==2.1.2
Werkzeug==2.3.7
sentence-transformers==2.2.2
numpy==1.24.3
